// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      lessons: {
        Row: {
          content_path: string
          created_at: string | null
          description: string | null
          id: string
          module_id: string
          order: number
          slug: string
          title: string
        }
        Insert: {
          content_path: string
          created_at?: string | null
          description?: string | null
          id?: string
          module_id: string
          order?: number
          slug: string
          title: string
        }
        Update: {
          content_path?: string
          created_at?: string | null
          description?: string | null
          id?: string
          module_id?: string
          order?: number
          slug?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "lessons_module_id_fkey"
            columns: ["module_id"]
            isOneToOne: false
            referencedRelation: "modules"
            referencedColumns: ["id"]
          },
        ]
      }
      modules: {
        Row: {
          created_at: string | null
          description: string | null
          id: string
          level: string
          order: number
          slug: string
          title: string
        }
        Insert: {
          created_at?: string | null
          description?: string | null
          id?: string
          level: string
          order?: number
          slug: string
          title: string
        }
        Update: {
          created_at?: string | null
          description?: string | null
          id?: string
          level?: string
          order?: number
          slug?: string
          title?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}